exercicios logicos 2

+++ errosDiversos/ProgramaComErroDeLogica01.java
@@ -9,7 +9,7 @@ public class ProgramaComErroDeLogica01
     public static void main(String args[]){
         int valor1 = 7;
         int valor2 = 8;
-        int media = valor1 + valor2 / 2;
+        double media = (valor1 + valor2) / 2;
         System.out.println("\fMédia Aritmética: " + media);
     }
 }
